test: fold the duplicate coverage the minimalism gate found into its one home
